adding VideoPacketTypeTag into evalvid-server model adding Tagging information to every packets sending from evalvid server applciation
