+++ model/evalvid-server.cc
@@ -20,8 +20,6 @@
  *
  */
 
-#include "ns3/tag.h"
-#include "ns3/packet.h"
 
 #include "ns3/log.h"
 #include "ns3/ipv4-address.h"
@@ -35,7 +33,7 @@
 #include "ns3/string.h"
 
 #include "evalvid-server.h"
-
+#include "video-type-tag.h"
 
 using namespace std;
 
@@ -363,84 +361,5 @@ EvalvidServer::HandleRead (Ptr<Socket> socket)
     }
 }
 
-/***************************************************************
- *           VideoPacketType Tags
- ***************************************************************/
-
-VideoPacketTypeTag::VideoPacketTypeTag ()
-  :Tag()
-{
-  NS_LOG_FUNCTION (this);
-}
-
-VideoPacketTypeTag::VideoPacketTypeTag (string videoPacketType)
-  :Tag(), m_videoPacketType(videoPacketType)
-{
-  NS_LOG_FUNCTION (this << videoPacketType);
-}
-
-void
-VideoPacketTypeTag::SetVideoPacketType (string n)
-{
-  NS_LOG_FUNCTION (this << n);
-  m_videoPacketType = n;
-}
-
-std::string
-VideoPacketTypeTag::GetVideoPacketType (void) const
-{
-  NS_LOG_FUNCTION (this);
-  return m_videoPacketType;
-}
-
-NS_OBJECT_ENSURE_REGISTERED (VideoPacketTypeTag);
-
-TypeId
-VideoPacketTypeTag::GetTypeId (void)
-{
-  static TypeId tid = TypeId ("ns3::VideoPacketTypeTag")
-    .SetParent<Tag> ()
-    .SetGroupName("Network")
-    .AddConstructor<VideoPacketTypeTag> ();
-  return tid;
-}
-TypeId
-VideoPacketTypeTag::GetInstanceTypeId (void) const
-{
-  return GetTypeId ();
-}
-uint32_t
-VideoPacketTypeTag::GetSerializedSize (void) const
-{
-  NS_LOG_FUNCTION (this);
-  uint32_t s = 1 + m_videoPacketType.size();  // +1 for name length field
-  return s;
-}
-void
-VideoPacketTypeTag::Serialize (TagBuffer i) const
-{
-  NS_LOG_FUNCTION (this << &i);
-  const char *n = m_videoPacketType.c_str();
-  uint8_t l = (uint8_t) m_videoPacketType.size ();
-
-  i.WriteU8 (l);
-  i.Write ( (uint8_t*) n , (uint32_t) l);
-}
-void
-VideoPacketTypeTag::Deserialize (TagBuffer i)
-{
-  NS_LOG_FUNCTION (this << &i);
-  uint8_t l = i.ReadU8();
-  char buf[256];
-
-  i.Read ( (uint8_t* ) buf, (uint32_t) l);
-  m_videoPacketType = std::string (buf, l);
-}
-void
-VideoPacketTypeTag::Print (std::ostream &os) const
-{
-  NS_LOG_FUNCTION (this << &os);
-  os << "VideoPacketType=" << m_videoPacketType;
-}
 
 } // Namespace ns3
